feat: adjust frame timestamp calculation
(cherry picked from commit b9f9c9fdd3d7125d2b2e911cd6e5209f9d06449d)

+++ teleop_client_lib/include/teleop_client/video_encoder.hpp
@@ -96,7 +96,7 @@ public:
 
 private:
     int64_t startTime_;
-    int frameIndex_;
+    int64_t frameIndex_;
     AVCodecContext *codecContext_;
     AVFrame *avFrame_;
     AVPacket *avPacket_;
